myProject/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-crispy-forms==2.0
django-storages==1.13.2
Pillow==9.5.0
requests==2.29.0
azure-core==1.26.3
azure-storage-blob==12.14.1
whitenoise==6.4.0
